+++ CHANGELOG.md
@@ -17,9 +17,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Introduce a trace-agent generator for Datadog Agent traces.
 - Stable throttle now has a 'timeout' configuration parameter to model IO done
   with timeout.
-- `logrotate_fs` now randomizes its read offset from the block cache in order to return unique contents per log file
+- `logrotate_fs` now randomizes its read offset from the block cache in order to
+  return unique contents per log file
 - Added a 'Datadog Intake' blackhole, supporting only the v2 metrics protobuf
   protocol for now. This captures metrics submitted into capture files.
+- Added the ability for lading to write capture files in parquet format.
 
 ## [0.28.0]
 ## Added